Refactor code and structure - renders screen
This new refactored application is a lot more modular and clearer. Also changed project structure to cmake (and clion).

It now supports multiple "shared" textures so that the HUD could be captured separately from the 3D framebuffer. It's also a lot more efficient and stable. The OpenXR loop is also implemented more properly and destruction of components is mostly implemented too now.

It's currently in a state where it's not as feature-rich as the very first D3D11 prototype, but it should be easy to implement those parts now.